feat: add user with jwt auth

+++ app/services/path_service.py
@@ -6,10 +6,11 @@ from fastapi import Depends
 from ortools.constraint_solver import pywrapcp, routing_enums_pb2
 
 from app.core.settings import settings
-from app.exceptions.erros import NotFoundError
+from app.exceptions.erros import ForbiddenError, NotFoundError
 from app.repositories.path_repository import PathRepository
 from app.schemas.filters_params_schema import SortEnum
 from app.schemas.path_schema import PathCreate, PathResponse, PathResponseList
+from app.schemas.user_schema import UserResponse
 
 
 class PathService:
@@ -19,27 +20,34 @@ class PathService:
     ) -> None:
         self.repository = repository
 
-    async def get_all_paths(
+    async def get_all_paths_by_user(
         self,
+        user: UserResponse,
         skip: int,
         limit: int,
         order_by: str,
         arranging: SortEnum,
     ) -> PathResponseList:
-        db_paths = await self.repository.search_all(
-            skip, limit, order_by, arranging
+        db_paths = await self.repository.search_all_by_user(
+            user.id, skip, limit, order_by, arranging
         )
         return PathResponseList(
             data=[PathResponse.model_validate(path) for path in db_paths]
         )
 
-    async def get_path_by_id(self, path_id: UUID) -> PathResponse:
+    async def get_path_by_id(
+        self, path_id: UUID, user: UserResponse
+    ) -> PathResponse:
         db_path = await self.repository.search(path_id)
         if db_path is None:
             raise NotFoundError
+        if db_path.user_id != user.id:
+            raise ForbiddenError
         return PathResponse.model_validate(db_path)
 
-    async def create_path(self, path: PathCreate) -> PathResponse:
+    async def create_path(
+        self, user: UserResponse, path: PathCreate
+    ) -> PathResponse:
         base_url = settings.OSRM_URL + 'table/v1/driving/'
         coords = [f'{path.pickup.lng},{path.pickup.lat}']
         coords.extend(f'{coord.lng},{coord.lat}' for coord in path.dropoff)
@@ -102,12 +110,15 @@ class PathService:
             dropoff.model_dump() if hasattr(dropoff, 'model_dump') else dropoff
             for dropoff in reordered_dropoffs
         ]
+        path_data['user_id'] = user.id
 
         db_path = await self.repository.create(path_data)
         return PathResponse.model_validate(db_path)
 
-    async def delete_path(self, path_id: UUID) -> None:
+    async def delete_path(self, path_id: UUID, user: UserResponse) -> None:
         path = await self.repository.search(path_id)
         if path is None:
             raise NotFoundError
+        if path.user_id != user.id:
+            raise ForbiddenError
         return await self.repository.delete(path)
